requirements.txt
twitchio==2.3.0
python-dotenv==0.20.0
requests==2.28.0

# Formatting
flake8==4.0.1
black==22.3.0